Extract repo session tree helpers
Move repo session path, tree, and session-info shaping into a focused helper module so the session DO stays centered on orchestration.

Made-with: Cursor
